Factor out TestResponse
